ability to customize docker storage driver
This commit allows a user to customize the Docker storage driver
by using the command line flag `--storage-driver=foo`, thus
setting the storage driver to use with the Docker daemon running
inside the parent Docker container.

If this command line argument is not passed in as the first option,
then the script will default to `overlay` per gh-56.